+++ changelog.md
@@ -163,6 +163,8 @@ its results in a DBF file that is read by the MLogit.R script. "RJDBC" is indeed
     - Add "font" properties to ScaleAndCompassLayer for label and scale fonts.
     - Use a fork specific version of openmap.jar that fixes a bug with EsriLayer.setModel(...).
     - Add legends in virtual network viewer.
+    - Replace 'importTables" property by 'import.tables'. Projects with the old property name are still accepted.
+    - Add the possibility to specify the colors to use for each mode in the statistics pie diagrams in the project file.
 
 - Breaking changes:
     - Simplify the API for modal split methods. This breaks existing plugin's. If an incompatible plugin is found, an error